add ability to change sync time

+++ feature_settings/src/main/java/dev/lkey/settings/presentation/viewmodel/SettingsEvent.kt
@@ -3,6 +3,6 @@ package dev.lkey.settings.presentation.viewmodel
 sealed interface SettingsEvent {
 
     data object OnLoadData: SettingsEvent
-    data class OnChooseSyncDuration(val duration: Long): SettingsEvent
+    data class OnChooseSyncDuration(val duration: Float): SettingsEvent
 
 }
